Backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
openai==0.27.0
boto3==1.18.0
python-dotenv==0.19.0
opencv-python==4.5.3.56
werkzeug==2.0.1
beautifulsoup4==4.10.0
requests==2.26.0
